vector-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
faiss-cpu==1.11.0.post1
sentence-transformers==2.4.0
numpy>=1.23,<2
